Remove an unnecessary directory level, and move witness into `internal`.
This is to avoid expanding our public API surface further than
necessary.
