+++ internal/witness/cmd/witness/Dockerfile
