+++ internal/witness/cmd/witness/example_config.yaml
