+++ internal/witness/cmd/witness/impl/witness.go
@@ -24,8 +24,8 @@ import (
 
 	"github.com/golang/glog"
 	ct "github.com/google/certificate-transparency-go"
-	ih "github.com/google/certificate-transparency-go/witness/witness/cmd/witness/internal/http"
-	"github.com/google/certificate-transparency-go/witness/witness/cmd/witness/internal/witness"
+	ih "github.com/google/certificate-transparency-go/internal/witness/cmd/witness/internal/http"
+	"github.com/google/certificate-transparency-go/internal/witness/cmd/witness/internal/witness"
 	"github.com/gorilla/mux"
 	_ "github.com/mattn/go-sqlite3" // Load drivers for sqlite3
 )
